Concurrently upgrade to 2.0
I was running into some trouble with npm 3.5.1 using concurrently (from /node_modules/.bin) it is referencing require('./lodash-mixins'), not present in the local .bin directory. An upgrade to concurrently 2.0.0 solves this, it also exposes the 'concurrently' binary, instead of 'concurrent', which should reduce the name clash with the concurrent npm package.